Remove unnecessary float to float conversions

+++ KillBox/src/Player.java
@@ -165,9 +165,9 @@ public class Player
 		for (int Point = 0; Point < MaxChecks; Point++)
 		{
 			// Increment bullet position
-			TravelX = TravelX * Step * (float)Math.cos((float)HorizontalAngle);
-			TravelY = TravelY * Step * (float)Math.sin((float)HorizontalAngle);
-			TravelZ = TravelZ * Step * (float)Math.sin((float)VerticalAngle);
+			TravelX = TravelX * Step * (float)Math.cos(HorizontalAngle);
+			TravelY = TravelY * Step * (float)Math.sin(HorizontalAngle);
+			TravelZ = TravelZ * Step * (float)Math.sin(VerticalAngle);
 
 			Player Hit = PointInPlayer(TravelX, TravelY, TravelZ);
